Various changes to generated code semantics.
Remove guaranteed uncurrying in favor of letting the compiler optimize depending
on a case by case.

Expose the external binding definitions over the interface file so trivial calls
to Javascript code are inlined.

Disable some warnings, on generated code, where it doesn't make sense.

Wrap the inner try expression in a function so the OCaml exception semantics are
preserved in the transformed code.